agno_tests/pyproject.toml
[project]
name = "agno-tests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.3.5",
    "anthropic>=0.79.0",
    "ddgs>=9.10.0",
    "fastapi>=0.115.0",
    "google-genai>=1.53.0",
    "google-generativeai>=0.8.5",
    "lancedb>=0.29.2",
    "openai>=2.16.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "sqlalchemy>=2.0.44",
    "yfinance>=1.1.0",
]
